Heap-allocate scalar type objects.
This fixes an incompatibility with the existing TensorFlow and JAX implementations of this module.

The existing implementations try not to register types like bfloat16 if they already exist. This is to avoid the situation where there are multiple different "bfloat16"s present in the same NumPy installation, which gets confusing fast. They do this by calling PyArray_TypeNumFromName() to find a type with tp_name "bfloat16". Before this change, the ml_dtypes type was named "ml_dtypes.bfloat16", which did not match, causing problems.

We can fix this by using a quirk of the Python interpreter. tp_name is only a fully-qualified name for statically allocated types. For heap-allocated types, it is unqualified, and the `__module__` field is used to determine the enclosing module (https://docs.python.org/3/c-api/typeobj.html#c.PyTypeObject.tp_name ).

PiperOrigin-RevId: 518544831