Explicitly enforce int4 bit representation with ignored high bits.
The previous implementation using a bit-field leaves the exact
representation implementation-defined, with some platforms storing
bit-fields packed left-to-right, others right-to-left, and the
masked bits unspecified.  This complicates serialization and
vectorized conversions.

With this change, we now explicitly store the value in the lower
4 bits, and leave the upper 4 bits unspecified.  The type is
constructed in such a way that correctness is preserved
regardless of the upper bit values.

PiperOrigin-RevId: 578735344