Add float8_base as friend class to FP4 and FP6 types.
This fixes a compile error on Clang 12 and below. Clang had a bug where if a superclass instantiated an instance of a subclass with a constructor defined in the superclass and brought into the subclass with a using-declaration, and the constructor had at least two arguments, compilation would fail. See https://godbolt.org/z/aeP9sP5x5 for an example. The error would complain that the superclass does not have access to the subclass's protected members, despite the constructor being declared in the superclass itself.

The fix is to make the superclass a friend class of the subclass. This is already done in float8.h, but wasn't done in mxfloat.h.

https://github.com/openxla/xla/pull/19096 was rolled back since it added an include of mxfloat.h in TensorFlow, causing an Android TensorFlow build using Android NDK r21e to fail since this NDK uses Clang 9.

PiperOrigin-RevId: 713847187