Initial complex32 and bcomplex32 implementation
This defines most things needed for complex32 and bcomplex32 as
well as related tests.

There will still be some things to do here, hopefully minor.
I am not sure about some of the header decisions and namings I did.
Also, as noted on the issue. The character codes are just random
right now.

Given the issue with `np.imag` and `np.real` I could also imagining
putting a warning on these dtypes for the moment.